Fix blockly datetime picker
